+++ certbot/CHANGELOG.md
@@ -94,7 +94,7 @@ More details about these changes can be found on our GitHub repo.
 
 ### Added
 
-
+*
 
 ### Changed