+++ certbot/CHANGELOG.md
@@ -6,7 +6,7 @@ Certbot adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
-*
+* Updated Windows installer to be signed and trusted in Windows 
 
 ### Changed
 
@@ -27,6 +27,7 @@ More details about these changes can be found on our GitHub repo.
 ### Added
 
 * Updated Apache/NGINX TLS configs to document contents are based on ssl-config.mozilla.org
+
 
 ### Changed